começo do client GraphQL
fazendo o commit logo pq minha placa de vídeo ta morrendo

+++ src/main/java/com/pedror/MusicaGraphQLService.java
@@ -49,20 +49,20 @@ public class MusicaGraphQLService {
 
   @Query
   @Description("Retorna todas as playlists de um usuário.")
-  public List<Playlist> getPlaylistsDoUsuario(Long idUsuario) {
-    return dados.getPlaylistsDoUsuario(idUsuario);
+  public List<Playlist> getPlaylistsDoUsuario(int idUsuario) {
+    return dados.getPlaylistsDoUsuario((long) idUsuario);
   }
 
   @Query
   @Description("Retorna todas as músicas de uma playlist.")
-  public List<Musica> getMusicasDaPlaylist(Long idPlaylist) {
-    return dados.getMusicasDaPlaylist(idPlaylist);
+  public List<Musica> getMusicasDaPlaylist(int idPlaylist) {
+    return dados.getMusicasDaPlaylist((long) idPlaylist);
   }
 
   @Query
   @Description("Retorna todas as playlists que contêm uma música.")
-  public List<Playlist> getPlaylistsComMusica(Long idMusica) {
-    return dados.getPlaylistsComMusica(idMusica);
+  public List<Playlist> getPlaylistsComMusica(int idMusica) {
+    return dados.getPlaylistsComMusica((long) idMusica);
   }
